Implementing better search functionality for devtools searchbar
Summary:
The Devtools searchbar is currently broken and only allows users to search for certain requests. This diff will fix this issue and also implement a new feature that allows users to search for multiple different requests at the same time, by adding spaces in between.

Ex: "searchA searchB" should search for requests with names that include searchA OR searchB

Reviewed By: tyao1

Differential Revision: D22162053

fbshipit-source-id: 360b75d4552a68c72aa9d5c8bb404bf60851cd74