Optimizing the process of sending store data in the devtools
Summary:
Our current, naive approach is to send the entire relay store data to the devtools on every update, but it’s inefficient. The goal of this task is to find a way to efficiently send only data that has been updated.

1. Get store events (specifically we want store.publish) so that we can get only the records that have been updated instead of the whole store.
2. Send these updates to the devtools store and add them to our map!

Results:
This optimization of the live store ended up being 8-9x faster!

Reviewed By: tyao1

Differential Revision: D22504029

fbshipit-source-id: 1387dd90140cf60f0b4273558ff30dd5b972a6c3