Update Snapshot Tab
Summary:
We currently have an empty snapshot tab and a refresh button that doesn't do anything.
Let's actually store snapshots in the snapshot tab and allow users to see them!
The snapshot button takes a snapshot of the store and stores it in the snapshot tab. Users can look back and see all their snapshots whenever they want.

Reviewed By: tyao1

Differential Revision: D22604825

fbshipit-source-id: 4313958ce56346f19d851f1523fe473be1344663