Bug: refreshing doesn't cause the devtools to reload
Summary:
A user pointed out a bug with the dev tools that refreshing the inspected page would not refresh the relay dev tools.

Upon inspection, I realized that we had just forgotten to add 'webNavigation' as a permission, which means that none of the chrome.WebNavigation functions were being called.

Reviewed By: tyao1

Differential Revision: D23063648

fbshipit-source-id: 104dbc471b2aea1768f3e64045e8478c29003d5b