RFC 1/2: Rename transactionID to networkRequestId in log events
Summary:
Our current log events for network events have a `transactionID` to identify the network request, however, it's not entirely clear /what/ transaction this is referring to, and e.g. if it includes processing the request, etc.

IMO, renaming this field to `networkRequestId` makes it a bit more clear what this value is.

If we agree with this naming, this commit updates devtools to support both `transactionID` or `networkRequestId` without breaking, and in the next diff I'll update the types in `RelayStoreTypes`.

Reviewed By: tyao1

Differential Revision: D28192255

fbshipit-source-id: e22b6233f53d5364a1b3aa921edb26b093c0d0c9