Fix bug in record list when when searching
Summary:
Fixing a bug where all the *record types* would continue to be visible even if nothing matched the searched string. Also clearing the selected record when starting a searching.

This should improve the overall experience, making it easier to spot the search results or when nothing in the store matches the search.

Reviewed By: tyao1

Differential Revision: D28627300

fbshipit-source-id: 3f21c96df4261daf230d965d90ab3fbd354a9e10